+++ asf_public_discourse_home_decarbonisation/utils/general_utils.py
@@ -1,0 +1,9 @@
+"""
+General utils.
+"""
+
+
+def list_chunks(orig_list: list, chunk_size: int = 100):
+    """Chunks list into batches of a specified chunk_size."""
+    for i in range(0, len(orig_list), chunk_size):
+        yield orig_list[i : i + chunk_size]
